+++ hack/generate-manifests.sh
@@ -13,19 +13,40 @@ for version in v1 v1beta1; do
   popd > /dev/null
 
   # Building Cluster API CRDs.
-  # The line `paths=sigs.k8s.io/cluster-api/api/v1alpha2`
-  # indicates that the CRD should contain only `v1alpha2`.
-  # At the time of writing, `paths=sigs.k8s.io/cluster-api/api` would
+  # At the time of writing, `paths=sigs.k8s.io/cluster-api/api/...` would
   # result in both `v1alpha2` and `v1alpha3` being included.
   ./tools/bin/controller-gen \
     crd \
-    paths=sigs.k8s.io/cluster-api/api/v1alpha2 \
+    paths="sigs.k8s.io/cluster-api/api/..." \
     output:dir="../config/crd/$version" \
     crd:crdVersions="$version"
 
-  # We only want Cluster and MachineDeployment for now, so delete the other two CAPI CRDs.
+  # Also build experimental Cluster API CRDs. Most importantly MachinePool CRD.
+  ./tools/bin/controller-gen \
+    crd \
+    paths="sigs.k8s.io/cluster-api/exp/api/..." \
+    output:dir="../config/crd/$version" \
+    crd:crdVersions="$version"
+
+  # We also want Provider specific CRDs for Azure.
+  ./tools/bin/controller-gen \
+    crd \
+    paths=sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3 \
+    output:dir="../config/crd/$version" \
+    crd:crdVersions="$version"
+
+  # With MachinePool related types.
+  ./tools/bin/controller-gen \
+    crd \
+    paths=sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3 \
+    output:dir="../config/crd/$version" \
+    crd:crdVersions="$version"
+
+  # Delete unused upstream CRDs.
   rm ../config/crd/$version/cluster.x-k8s.io_machines.yaml
   rm ../config/crd/$version/cluster.x-k8s.io_machinesets.yaml
+  rm ../config/crd/$version/cluster.x-k8s.io_machinehealthchecks.yaml
+  rm ../config/crd/$version/infrastructure.cluster.x-k8s.io_azuremachinetemplates.yaml
 
   # Add .metadata.name validation to Release CRD using kustomize since
   # kubebuilder comments can't modify metav1.ObjectMeta