+++ pkg/apis/application/v1alpha1/app_types.go
@@ -166,6 +166,8 @@ type AppStatus struct {
 }
 
 type AppStatusRelease struct {
+	// +kubebuilder:validation:Optional
+	// +nullable
 	// LastDeployed is the time when the app was last deployed.
 	LastDeployed metav1.Time `json:"lastDeployed,omitempty"`
 	// Reason is the description of the last status of helm release when the app is